Fix to avoid CMake 2.8.11 spurious error
CMake 2.8.11 has a check that fails because it sees a relative "include"
dir as target_include_directories.
However it is in a conditional expression (INSTALL_INTERFACE) that is
supposed to prepend the install prefix (thus no relative path in the
end).
The check still fails, probably an internal error in 2.8.11 (no problem
for any version above that).

To fix it, the install prefix is prepended manually.
It is refered to using a generator expression instead of
CMAKE_INSTALL_PREFIX : with the CMAKE_... variable it changes the
output, not with the generator expression.

Updated doc.