Fix interface include dir for absolute CMAKE_INSTALL_INCLUDEDIR
If the user-specified CMAKE_INSTALL_INCLUDEDIR is an absolute path (as
permitted by GNUInstallDirs), joining it to the install prefix is not
correct. In fact, we don't need to specify the install prefix here at
all, as a relative interface install include dir is already assumed to
be relative to the install prefix.

https://cmake.org/cmake/help/latest/prop_tgt/INTERFACE_INCLUDE_DIRECTORIES.html